chore: remove old ticket claim page
